fix(windows,network): stabilize full build TLS and timeout tests
Windows full-build fixes:

* Include windows.h before bcrypt.h in rt_crypto_module so the Windows SDK defines NTSTATUS, ULONG, LPCWSTR, and related types before bcrypt declarations are parsed. Clamp BCryptGenRandom chunks with UINT32_MAX, which is the portable bound for the ULONG argument without depending on DWORD_MAX.

* Teach the Windows TLS CertificateVerify path to verify ecdsa_secp256r1_sha256 signatures with CNG. TLS sends ECDSA signatures as DER SEQUENCE { r, s }, while BCryptVerifySignature expects raw fixed-width r || s bytes, so the Windows path now decodes and pads the scalars before importing the certificate public key with CryptImportPublicKeyInfoEx2.

* Relax RTNetworkTests timeout assertions on Windows so accept/recv timeout checks still reject immediate returns but tolerate normal Windows timer granularity and suite-load jitter. The full test suite was intermittently measuring a 100 ms timeout below the old 90 ms lower bound.

Verification:

* cmake --build build --config Debug --target viper_rt_network_obj test_rt_network_highlevel -j %NUMBER_OF_PROCESSORS%

* ctest --test-dir build -C Debug --output-on-failure -R ^test_rt_network_highlevel$

* ctest --test-dir build -C Debug --output-on-failure -R ^(test_rt_crypto|test_rt_network|test_rt_network_timeout|test_rt_network_harden|test_rt_tls_cert|test_rt_tls_wrapper|test_rt_network_highlevel)$

* cmd /c scripts\build_viper.cmd

+++ src/runtime/network/rt_crypto_module.c
@@ -23,8 +23,8 @@ extern void arc4random_buf(void *buf, size_t nbytes);
 #ifndef WIN32_LEAN_AND_MEAN
 #define WIN32_LEAN_AND_MEAN
 #endif
-#include <bcrypt.h>
 #include <windows.h>
+#include <bcrypt.h>
 #ifndef NT_SUCCESS
 #define NT_SUCCESS(Status) (((NTSTATUS)(Status)) >= 0)
 #endif
@@ -62,8 +62,8 @@ static int module_os_entropy(uint8_t *buf, size_t len) {
     size_t off = 0;
     while (off < len) {
         size_t chunk = len - off;
-        if (chunk > ULONG_MAX)
-            chunk = ULONG_MAX;
+        if (chunk > UINT32_MAX)
+            chunk = UINT32_MAX;
         NTSTATUS status =
             BCryptGenRandom(NULL, buf + off, (ULONG)chunk, BCRYPT_USE_SYSTEM_PREFERRED_RNG);
         if (!NT_SUCCESS(status))
